MOM-249 restructruring
Move api calls to dedicated "api wrapper" class
Add responsible column in revision front end
